requirements.txt
setuptools
scipy==1.12.0
numpy==1.26.4
pandas==2.2.0
matplotlib
Cython==3.0.8
lru-dict
mpmath==1.3.0
tqdm
scikit-learn==1.4.1.post1
bigtree
lru-dict
ray==2.39.0